requirements.txt
tensorflow-cpu
numpy
geopandas
shapely
rasterio
Pillow
tqdm